+++ src/modules/location-huds/locations/cerulean-skyport/index.js
@@ -2,6 +2,7 @@ import { addHudStyles } from '@utils';
 
 import { addAirshipRandomizer } from '../../shared/airship-randomizer';
 import { addHullClick } from './ledger';
+import { initCurrentRaidIndicator } from './current-raid';
 import { initRaidFavorites } from './raid-favorites';
 
 import fullWidthAirshipStyles from '../floating-islands/full-width-airship.css';
@@ -15,4 +16,5 @@ export default async () => {
   addAirshipRandomizer();
   addHullClick();
   initRaidFavorites();
+  initCurrentRaidIndicator();
 };